06_Data_Version_Control/pyproject.toml
[project]
name = "emlo4-session-04-ajithvcoder"
version = "0.1.0"
description = "Use lightning to train, eval and infer a dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.8.0",
    "dvc-gdrive>=3.0.1",
    "dvc>=3.55.2",
    "dvc[gs]",
    "lightning[extra]==2.4.0",
    "loguru>=0.7.2",
    "rich>=13.8.1",
    "tensorboard==2.17.1",
    "timm==1.0.9",
    "torch==2.4.1+cpu",
    "torchvision==0.19.1+cpu",
    "gdown==5.2.0",
    "hydra-core==1.3.0",
    "pandas==2.2.3",
    "scikit-learn==1.5.2",
    "comet_ml==3.47.0",
    "python-dotenv==1.0.1",
    "matplotlib==3.9.2",
    "rootutils==1.0.7",
    "pytest==8.3.3",
    "pytest-cov==5.0.0",
    "coverage==7.6.3",
    "pytest-order==1.3.0",
    "pytest-ordering==0.6",
    "pytest-dependency==0.6.0"
]

[tool.coverage.report]
include = ["src/**/*.py"]
show_missing = true